// libs/duality-social-lib/src/lib/interfaces/default-reactions.ts
import { IDefaultReaction } from './default-reaction.ts';

export interface IDefaultReactions {
  Angry: IDefaultReaction;
  Care: IDefaultReaction;
  Celebrate: IDefaultReaction;
  Hug: IDefaultReaction;
  'Huh?': IDefaultReaction;
  Laugh: IDefaultReaction;
  Like: IDefaultReaction;
  Love: IDefaultReaction;
  Sad: IDefaultReaction;
  Wow: IDefaultReaction;
  Yuck: IDefaultReaction;
}